Average results over valid days

This function computes, using valid days only, the mean of each of the metrics obtained using the recap_by_day function.

Usage

average_results(data, minimum_wear_time = 10)

Arguments

data

A dataframe obtained using the prepare_dataset, mark_wear_time, mark_intensity and then recap_by_day functions.

minimum_wear_time

A numeric value (in hours) to set the minimum wear time duration for validating a day.

Value

A dataframe.

Examples

file <- system.file("extdata", "acc.agd", package = "activAnalyzer")
mydata <- prepare_dataset(data = file, epoch_len_out = 60, col_time_stamp = "timestamp")
mydata_with_wear_marks <- mark_wear_time(
    dataset = mydata, 
    TS = "timestamp", 
    cts  = "vm", frame = 90, 
    allowanceFrame = 2, 
    streamFrame = 30
    )
#> frame is 90
#> streamFrame is 30
#> allowanceFrame is 2
mydata_with_intensity_marks <- mark_intensity(
    data = mydata_with_wear_marks, 
    col_axis = "vm", 
    equation = "Sasaki et al. (2011) [Adults]",
    sed_cutpoint = 200, 
    mpa_cutpoint = 2690, 
    vpa_cutpoint = 6167, 
    age = 32,
    weight = 67,
    sex = "male",
    col_steps = "steps"
    )
#> You have computed BMR using the following inputs: 
#>     age = 32
#>     weight = 67
#>     sex = male
#> You have computed METs using the following inputs: 
#>     equation = Sasaki et al. (2011) [Adults]
#>     weight = 67
#>     sex = male
#> You have computed intensity metrics using the following inputs: 
#>     axis = vm
#>     sed_cutpoint = 200
#>     mpa_cutpoint = 2690
#>     vpa_cutpoint = 6167
#>     equation = Sasaki et al. (2011) [Adults]
#>     age = 32
#>     weight = 67
#>     sex = male
summary_by_day <- recap_by_day(
    data = mydata_with_intensity_marks, 
    age = 40, weight = 70, 
    sex = "male"
    )
#> You have computed BMR using the following inputs: 
#>     age = 40
#>     weight = 70
#>     sex = male
average_results(data = summary_by_day, minimum_wear_time = 10)
#> # A tibble: 1 × 29
#>   valid_days wear_time total_counts_axis1 total_counts_vm total_steps
#>        <int>     <dbl>              <dbl>           <dbl>       <dbl>
#> 1          5      768.            513109.         951483.       14869
#> # … with 24 more variables: total_kcal_wear_time <dbl>, minutes_SED <dbl>,
#> #   minutes_LPA <dbl>, minutes_MPA <dbl>, minutes_VPA <dbl>,
#> #   minutes_MVPA <dbl>, percent_SED <dbl>, percent_LPA <dbl>,
#> #   percent_MPA <dbl>, percent_VPA <dbl>, percent_MVPA <dbl>,
#> #   max_steps_60min <dbl>, max_steps_30min <dbl>, max_steps_20min <dbl>,
#> #   max_steps_5min <dbl>, max_steps_1min <dbl>, peak_steps_60min <dbl>,
#> #   peak_steps_30min <dbl>, peak_steps_20min <dbl>, peak_steps_5min <dbl>, …