Get intensity gradient values and graphics

The values and graphics are respectively computed and created from the daily periods set for analysis and are based on detected wear time.

Usage

get_ig_results(
  data,
  col_axis = "vm",
  col_time = "time",
  valid_wear_time_start = "00:00:00",
  valid_wear_time_end = "23:59:59",
  start_first_bin = 0,
  start_last_bin = 10000,
  bin_width = 500,
  cor_factor = 1
)

Arguments

data

A dataframe obtained using the prepare_dataset, mark_wear_time, and then the mark_intensity functions. Data should be grouped by day and then nested.

col_axis

A character value to indicate the name of the variable to be used to compute total time per bin of intensity.

col_time

A character value to indicate the name of the variable to be used to determine the epoch length of the dataset.

valid_wear_time_start

A character value with the HH:MM:SS format to set the start of the daily period that will be considered for computing metrics.

valid_wear_time_end

A character value with the HH:MM:SS format to set the end of the daily period that will be considered for computing metrics.

start_first_bin

A numeric value to set the lower bound of the first bin of the intensity band (in counts/epoch duration).

start_last_bin

A numeric value to set the lower bound of the last bin of the intensity band (in counts/epoch duration).

bin_width

A numeric value to set the width of the bins of the intensity band (in counts/epoch duration).

cor_factor

A numeric value resulting from the ratio between 60s and the epoch length of the analysed dataset. This is used to convert the number of rows into minutes when getting the results.

Value

A list of objects.