Summarize results by day

This function summarizes accelerometer results for each day of the measurement period.

recap_by_day(
  data,
  col_date = "date",
  age = 40,
  weight = 70,
  sex = c("male", "female", "undefined")
)

Arguments

data

A dataframe obtained using the prepare_dataset, mark_wear_time, and then mark_intensity functions.

col_date

A character value to indicate the name of the date variable.

age

A numeric value in yr.

weight

A numeric value in kg.

sex

A character value.

Value

A dataframe.

Details

The following metrics are computed:

wear_time: total wear time

total_counts_axis1: total counts for the vertical axis

total_counts_vm: total counts for the vector magnitude

total_steps: total step count

total_kcal_wear_time: total kilocalories

minutes_SED: total minutes spent in SED behavior

minutes_LPA: total minutes spent in LPA behavior

minutes_MPA: total minutes spent in MPA behavior

minutes_VPA: total minutes spent in VPA behavior

minutes_MVPA: total minutes spent in MVPA behavior

percent_SED: proportion of wear time spent in SED behavior

percent_LPA: proportion of wear time spent in LPA behavior

percent_MPA: proportion of wear time spent in MPA behavior

percent_VPA: proportion of wear time spent in VPA behavior

percent_MVPA: proportion of wear time spent in MPVA behavior

max_steps_60min: best step accumulation per minute averaged over a window of 60 continuous minutes

max_steps_30min: best step accumulation per minute averaged over a window of 30 continuous minutes

max_steps_20min: best step accumulation per minute averaged over a window of 20 continuous minutes

max_steps_5min: best step accumulation per minute averaged over a window of 5 continuous minutes

max_steps_1min: best step accumulation per minute over a window of 1 minute

peak_steps_60min: step accumulation per minute averaged over the best 60 continuous or discontinuous minutes

peak_steps_30min: step accumulation per minute averaged over the best 30 continuous or discontinuous minutes

peak_steps_20min: step accumulation per minute averaged over the best 20 continuous or discontinuous minutes

peak_steps_5min: step accumulation per minute averaged over the best 5 continuous or discontinuous minutes

peak_steps_1min: step accumulation per minute over the best minute (same result as for max_steps_1min)

mets_hours_mvpa: total MET-hours spent during MPVA behavior

ratio_mvpa_sed: ratio between MVPA and SED times (minutes_MVPA / minutes_SED)

PAL is computed as follows: total energy expenditure (TEE) is divided by BMR. TEE is obtained by summing the kilocalories measured during wear time epochs and the kilocalories likely expended during nonwear time epochs (that is, kilocalories associated to BMR, as it is assumed that the device was mainly not worn during sleeping periods if any, periods during which energy expenditure is near of BMR), and by multiplying this sum by 10/9 to take into account the thermic effect of food. Of course, such calculations may conduct to underestimate TEE and PAL if the device was removed during prolonged periods of physical activity. Moreover, even if the device was correctly worn, the estimate of PAL is very approximate since both BMR and kilocalories related to wear time are estimated using methods that may not be accurate at the individual level.

Examples

file <- system.file("extdata", "acc.agd", package = "activAnalyzer")
mydata <- prepare_dataset(data = file, epoch_len_out = 60, col_time_stamp = "timestamp")
mydata_with_wear_marks <- mark_wear_time(
    dataset = mydata, 
    TS = "timestamp", 
    cts  = "vm", 
    frame = 90, 
    allowanceFrame = 2, 
    streamFrame = 30
    )
#> frame is 90
#> streamFrame is 30
#> allowanceFrame is 2
mydata_with_intensity_marks <- mark_intensity(
    data = mydata_with_wear_marks, 
    col_axis = "vm", 
    equation = "Sasaki et al. (2011) [Adults]",
    sed_cutpoint = 200, 
    mpa_cutpoint = 2690, 
    vpa_cutpoint = 6167, 
    age = 32,
    weight = 67,
    sex = "male",
    col_steps = "steps"
    )
#> You have computed BMR using the following inputs: 
#>     age = 32
#>     weight = 67
#>     sex = male
#> You have computed METs using the following inputs: 
#>     equation = Sasaki et al. (2011) [Adults]
#>     weight = 67
#>     sex = male
#> You have computed intensity metrics using the following inputs: 
#>     axis = vm
#>     sed_cutpoint = 200
#>     mpa_cutpoint = 2690
#>     vpa_cutpoint = 6167
#>     equation = Sasaki et al. (2011) [Adults]
#>     age = 32
#>     weight = 67
#>     sex = male
recap_by_day(
    data = mydata_with_intensity_marks, 
    age = 40, 
    weight = 70, 
    sex = "male"
    )
#> You have computed BMR using the following inputs: 
#>     age = 40
#>     weight = 70
#>     sex = male
#> # A tibble: 6 × 29
#>   date   wear_time total_counts_ax… total_counts_vm total_steps total_kcal_wear…
#>   <chr>      <dbl>            <int>           <dbl>       <int>            <dbl>
#> 1 2021-…       720           359125         706252.       14056            1264.
#> 2 2021-…       811           495257         923432.       14595            1528.
#> 3 2021-…       718          1222315        1930805.       21635            2399.
#> 4 2021-…       770           320847         781469.       13744            1340.
#> 5 2021-…       820           167999         415457.       10315            1074.
#> 6 2021-…         0                0              0            0               0 
#> # … with 23 more variables: minutes_SED <dbl>, minutes_LPA <dbl>,
#> #   minutes_MPA <dbl>, minutes_VPA <dbl>, minutes_MVPA <dbl>,
#> #   percent_SED <dbl>, percent_LPA <dbl>, percent_MPA <dbl>, percent_VPA <dbl>,
#> #   percent_MVPA <dbl>, max_steps_60min <dbl>, max_steps_30min <dbl>,
#> #   max_steps_20min <dbl>, max_steps_5min <dbl>, max_steps_1min <dbl>,
#> #   peak_steps_60min <dbl>, peak_steps_30min <dbl>, peak_steps_20min <dbl>,
#> #   peak_steps_5min <dbl>, peak_steps_1min <dbl>, mets_hours_mvpa <dbl>, …